Use `parLapply()` instead of `mclapply()`
This should make it possible to run parts of the simulation in
parallel also on Windows, although this has not been tested yet.  For
now, using multiple cores only helps when there is a large number of
species in the simulation.  parLapply() seems to have more reasonable
memory requirements than forking with mclapply(), although it can take
little while to set up all the workers at the start, possibly
depending on the simulation as well.