Ecology loop reimplementation
It now uses lapply() to iterate through sites and should spend less
time updating species abundances in the `all_species` list,
particularly when there are thousands of species in many sites.  It
accomplishes this by creating an abundance matrix of the occupied
sites and updating the whole abundance vector of each species instead
of updating each element of the abundance vector (each site)
individually.

The parallelization of sites with parLapply() is a first attempt and
it probably should not be used as it is in this commit, as it is even
more inefficient than using lapply() (at least in some cases) .