Create a separate function to get data for the ecology simulation
The function get_site_info() takes a site identifier and some objects
from which it can retrieve the data that is required by the
apply_ecology() function, and returns a list with the data for that
site.  Getting this data requires access to objects which can get very
large, and it seems that it is costly to pass these objects to each of
the worker processes if the update_site_abundances() function is run
in parallel (multiple sites being processed simultaneously).
Therefore, the code that gets the necessary data was taken out of the
update_site_abundances() function.

Specifically, note that getting the trait data can take a significant
amount of time if there are many sites with very high species
richness.

This change should allow to take advantage of parallelization if the
user supplied apply_ecology() function is computationally demanding.
However, the function that gets all the data for the ecology
simulation is not run in parallel for the reasons mentioned above.  It
would be good to find a way to retrieve this data (species traits in
particular) in a more efficient manner.