Do not assume that apply_ecology() returns a named abundance vector
Although the abundance vector that is passed to the apply_ecology()
function has the species IDs as names, this user defined function
might not return a named vector, so it is a mistake to try to use the
names of the returned vector as indices (after trying to convert them
to integers).  However, it is assumed that the user provided function
keeps the same species order as the provided abundance vector.