Reimplement state saving
In the documentation of run_simulation(), it says that it is possible
to save the internal state to restart the simulation later, but this
threw an error because the directory to save the objects was never
created.

Additionally, specifying a numeric vector with the time steps to save
did not work, since this was not implemented.

This commit splits the original "save_state" parameter into two: now
"save_state" specifies in which time steps to create a save of the
simulation state and "last_state_only" determines whether to keep all
saves or only the last one (the default is keeping only the last).

Also, the second parameter of save_val() had the same name as the
function.  That was changed.