Restore state properly when using a config object instead of a file
Config objects have now a mandatory `gen3sis$general$config_name`
variable that is set automatically from the file name when they are
generated from a file.  In theory, this variable could also be set
explicitly in the config file, and it would override the default of
using the file name.

This variable contains the name of the subdirectory where the output
of the simulation associated to the config is saved.